python/requirements.txt
tensorflow
keras
matplotlib
efficientnet
numpy
pandas
scikit-learn
Flask
Flask-Cors
Pillow